requirements.txt
altair
pandas
streamlit
google-cloud-bigquery==3.19.0
google-auth==2.29.0